TEST: Fix tests for PY312, NP2

+++ nitransforms/io/afni.py
@@ -237,7 +237,7 @@ def _is_oblique(affine, thres=OBLIQUITY_THRESHOLD_DEG):
     True
 
     """
-    return (obliquity(affine).max() * 180 / pi) > thres
+    return float(obliquity(affine).max() * 180 / pi) > thres
 
 
 def _afni_deobliqued_grid(oblique, shape):
